+++ R/mixgb.R
@@ -39,6 +39,7 @@
 #' }
 #' @param save.models Whether to save imputation models for imputing new data later on. Default: \code{FALSE}
 #' @param save.vars For the purpose of imputing new data, the imputation models for response variables specified in \code{save.vars} will be saved. The values in \code{save.vars} can be a vector of names or indices. By default, only the imputation models for variables with missing values in the original data will be saved (\code{save.vars = NULL}). To save imputation models for all variables, users can specify it with \code{save.vars = colnames(data)}.
+#' @param save.models.folder Users can specify a folder directory to save all imputation models. Models will be saved as JSON format by internally calling \code{xgb.save()}, which is recommended by XGBoost.
 #' @param verbose Verbose setting for mixgb. If \code{TRUE}, will print out the progress of imputation. Default: \code{FALSE}.
 #' @param xgb.params A list of XGBoost parameters. For more details, please check \href{https://xgboost.readthedocs.io/en/stable/parameter.html}{XGBoost documentation on parameters}.
 #' @param nrounds The maximum number of boosting iterations for XGBoost. Default: 100
@@ -58,7 +59,8 @@
 mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = FALSE,
                   pmm.type = "auto", pmm.k = 5, pmm.link = "prob",
                   initial.num = "normal", initial.int = "mode", initial.fac = "mode",
-                  save.models = FALSE, save.vars = NULL, verbose = F,
+                  save.models = FALSE, save.vars = NULL, save.models.folder = NULL,
+                  verbose = F,
                   xgb.params = list(),
                   nrounds = 100, early_stopping_rounds = 10, print_every_n = 10L, xgboost_verbose = 0, ...) {
   if (!(is.data.frame(data) || is.matrix(data))) {
@@ -68,6 +70,19 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
   if (!is.data.table(data)) {
     data <- as.data.table(data)
   }
+
+  # check whether to use xgb.save()
+  if (!is.null(save.models.folder)) {
+    if (!dir.exists(save.models.folder)) {
+      dir.create(file.path(save.models.folder), recursive = TRUE)
+    }
+    save.models <- TRUE
+    XGB.save <- TRUE
+  } else {
+    XGB.save <- FALSE
+  }
+
+
 
   xgb.params <- do.call("default_params", xgb.params)
 
@@ -79,8 +94,7 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
       warning(" `ordinalAsInteger` will be coerced to FALSE when `pmm.type = NULL`")
     }
     ord.fac <- names(Filter(is.ordered, data))
-    # ord.fac<- colnames(data)[sapply(data,is.ordered)]
-    ## data[,c(ord.fac) := lapply(.SD, as.integer), .SDcols = ord.fac]
+
     if (length(ord.fac) > 0) {
       data[, c(ord.fac) := lapply(.SD, fac2int), .SDcols = ord.fac]
     }
@@ -104,7 +118,6 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
   sorted.types <- initial.obj$sorted.types
   sorted.names <- initial.obj$sorted.names
   sorted.naSums <- initial.obj$sorted.naSums
-  ##
   sorted.idx <- initial.obj$sorted.idx
 
   missing.idx <- initial.obj$missing.idx
@@ -220,7 +233,6 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
     # return(mixgb.obj)
   } else {
     # save.models=TRUE, save models for imputing new data..........................................................................................
-
     # save params of this imputer for impute.new().....................................................
     params <- list()
     params$initial.num <- initial.num
@@ -245,7 +257,14 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
 
 
     # pre-allocation for saved models: for each of m, save Nmodels
-    XGB.models <- vector("list", m)
+    if (isTRUE(XGB.save)) {
+      # save the model dir
+      XGB.models <- vector("list", m)
+    } else {
+      # save the model
+      XGB.models <- vector("list", m)
+    }
+
 
     #--------------------------------------------------------
     if (bootstrap == FALSE) {
@@ -272,13 +291,25 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
           }
         }
         # the last iteration, update the imputed dataset, also save models
-        saved.obj <- mixgb_save(
-          save.vars = save.vars, save.p = save.p, extra.vars = extra.vars, extra.types = extra.types, pmm.type = pmm.type, pmm.link = pmm.link, pmm.k = pmm.k,
-          yobs.list = yobs.list, yhatobs.list = yhatobs.list, sorted.dt = sorted.dt, missing.vars = missing.vars, sorted.names = sorted.names,
-          Na.idx = Na.idx, missing.types = missing.types, Ncol = Ncol,
-          xgb.params = xgb.params,
-          nrounds = nrounds, early_stopping_rounds = early_stopping_rounds, print_every_n = print_every_n, verbose = xgboost_verbose, ...
-        )
+        if (isTRUE(XGB.save)) {
+          saved.obj <- mixgb_xgb_save(
+            save.models.folder = save.models.folder, i = i,
+            save.vars = save.vars, save.p = save.p, extra.vars = extra.vars, extra.types = extra.types, pmm.type = pmm.type, pmm.link = pmm.link, pmm.k = pmm.k,
+            yobs.list = yobs.list, yhatobs.list = yhatobs.list, sorted.dt = sorted.dt, missing.vars = missing.vars, sorted.names = sorted.names,
+            Na.idx = Na.idx, missing.types = missing.types, Ncol = Ncol,
+            xgb.params = xgb.params,
+            nrounds = nrounds, early_stopping_rounds = early_stopping_rounds, print_every_n = print_every_n, verbose = xgboost_verbose, ...
+          )
+        } else {
+          saved.obj <- mixgb_save(
+            save.vars = save.vars, save.p = save.p, extra.vars = extra.vars, extra.types = extra.types, pmm.type = pmm.type, pmm.link = pmm.link, pmm.k = pmm.k,
+            yobs.list = yobs.list, yhatobs.list = yhatobs.list, sorted.dt = sorted.dt, missing.vars = missing.vars, sorted.names = sorted.names,
+            Na.idx = Na.idx, missing.types = missing.types, Ncol = Ncol,
+            xgb.params = xgb.params,
+            nrounds = nrounds, early_stopping_rounds = early_stopping_rounds, print_every_n = print_every_n, verbose = xgboost_verbose, ...
+          )
+        }
+
         # if pmm.type=NULL
         imputed.data[[i]] <- saved.obj$sorted.dt[, origin.names, with = FALSE]
         XGB.models[[i]] <- saved.obj$xgb.models
@@ -312,13 +343,25 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
             )
           }
         }
-        saved.obj <- mixgb_bootsave(
-          BNa.idx = boot.result$BNa.idx, boot.dt = boot.result$boot.dt, save.vars = save.vars, save.p = save.p, extra.vars = extra.vars, extra.types = extra.types, pmm.type = pmm.type, pmm.link = pmm.link, pmm.k = pmm.k,
-          yobs.list = yobs.list, yhatobs.list = yhatobs.list, sorted.dt = sorted.dt, missing.vars = missing.vars, sorted.names = sorted.names,
-          Na.idx = Na.idx, missing.types = missing.types, Ncol = Ncol,
-          xgb.params = xgb.params,
-          nrounds = nrounds, early_stopping_rounds = early_stopping_rounds, print_every_n = print_every_n, verbose = xgboost_verbose, ...
-        )
+        if (isTRUE(XGB.save)) {
+          saved.obj <- mixgb_boot_xgb_save(
+            save.models.folder = save.models.folder, i = i,
+            BNa.idx = boot.result$BNa.idx, boot.dt = boot.result$boot.dt, save.vars = save.vars, save.p = save.p, extra.vars = extra.vars, extra.types = extra.types, pmm.type = pmm.type, pmm.link = pmm.link, pmm.k = pmm.k,
+            yobs.list = yobs.list, yhatobs.list = yhatobs.list, sorted.dt = sorted.dt, missing.vars = missing.vars, sorted.names = sorted.names,
+            Na.idx = Na.idx, missing.types = missing.types, Ncol = Ncol,
+            xgb.params = xgb.params,
+            nrounds = nrounds, early_stopping_rounds = early_stopping_rounds, print_every_n = print_every_n, verbose = xgboost_verbose, ...
+          )
+        } else {
+          saved.obj <- mixgb_bootsave(
+            BNa.idx = boot.result$BNa.idx, boot.dt = boot.result$boot.dt, save.vars = save.vars, save.p = save.p, extra.vars = extra.vars, extra.types = extra.types, pmm.type = pmm.type, pmm.link = pmm.link, pmm.k = pmm.k,
+            yobs.list = yobs.list, yhatobs.list = yhatobs.list, sorted.dt = sorted.dt, missing.vars = missing.vars, sorted.names = sorted.names,
+            Na.idx = Na.idx, missing.types = missing.types, Ncol = Ncol,
+            xgb.params = xgb.params,
+            nrounds = nrounds, early_stopping_rounds = early_stopping_rounds, print_every_n = print_every_n, verbose = xgboost_verbose, ...
+          )
+        }
+
 
         # if pmm.type=NULL
         imputed.data[[i]] <- saved.obj$sorted.dt[, origin.names, with = FALSE]
@@ -345,7 +388,7 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
     if (verbose) {
       cat("\n")
     }
-    mixgb.obj <- list("imputed.data" = imputed.data, "XGB.models" = XGB.models, "params" = params)
+    mixgb.obj <- list("imputed.data" = imputed.data, "XGB.models" = XGB.models, "params" = params, "XGB.save" = XGB.save)
     class(mixgb.obj) <- "mixgbObj"
     return(mixgb.obj)
   }
@@ -355,6 +398,8 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
 
 #' Auxiliary function for validating xgb.params
 #' @description Auxiliary function for setting up the default XGBoost-related hyperparameters for mixgb and checking the \code{xgb.params} argument in \code{mixgb()}. For more details on XGBoost hyperparameters, please refer to \href{https://xgboost.readthedocs.io/en/stable/parameter.html}{XGBoost documentation on parameters}.
+#' @param device Can be either "cpu" or "cuda". For ther options please refer to \href{https://xgboost.readthedocs.io/en/stable/parameter.html#general-parameters}{XGBoost documentation on parameters}.
+#' @param tree_method Options: "auto", "exact", "approx", and "hist". Default: "hist".
 #' @param eta Step size shrinkage. Default: 0.3.
 #' @param gamma Minimum loss reduction required to make a further partition on a leaf node of the tree. Default: 0
 #' @param max_depth Maximum depth of a tree. Default: 3.
@@ -367,12 +412,9 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
 #' @param colsample_bynode Subsampling ratio of columns for each node. Default: 1.
 #' @param lambda L2 regularization term on weights. Default: 1.
 #' @param alpha L1 regularization term on weights. Default: 0.
-#' @param tree_method Options: "auto", "exact", "approx", "hist", and "gpu_hist". Default: "auto".
 #' @param max_leaves Maximum number of nodes to be added (Not used when \code{tree_method = "exact"}. Default: 0.
 #' @param max_bin Maximum number of discrete bins to bucket continuous features (Only used when \code{tree_method} is either \code{hist}, \code{approx} or \code{gpu_hist}. Default: 256.
-#' @param predictor Default: "auto"
 #' @param num_parallel_tree The number of parallel trees used for boosted random forests. Default: 1.
-#' @param gpu_id Which GPU device should be used. Default: 0.
 #' @param nthread The number of CPU threads to be used. Default: -1 (all available threads).
 #' @return A list of hyperparameters.
 #' @export
@@ -384,6 +426,18 @@ mixgb <- function(data, m = 5, maxit = 1, ordinalAsInteger = FALSE, bootstrap = 
 #'
 #' xgb.params <- do.call("default_params", xgb.params)
 #' xgb.params
-default_params <- function(eta = 0.3, gamma = 0, max_depth = 3, min_child_weight = 1, max_delta_step = 0, subsample = 0.7, sampling_method = "uniform", colsample_bytree = 1, colsample_bylevel = 1, colsample_bynode = 1, lambda = 1, alpha = 0, tree_method = "auto", max_leaves = 0, max_bin = 256, predictor = "auto", num_parallel_tree = 1, gpu_id = 0, nthread = -1) {
-  list(eta = eta, gamma = gamma, max_depth = max_depth, min_child_weight = min_child_weight, subsample = subsample, sampling_method = sampling_method, colsample_bytree = colsample_bytree, colsample_bylevel = colsample_bylevel, colsample_bynode = colsample_bynode, lambda = lambda, alpha = alpha, tree_method = tree_method, max_leaves = max_leaves, max_bin = max_bin, predictor = predictor, num_parallel_tree = num_parallel_tree, gpu_id = gpu_id, nthread = nthread)
+default_params <- function(device = "cpu", tree_method = "hist", eta = 0.3, gamma = 0, max_depth = 3, min_child_weight = 1, max_delta_step = 0,
+                           subsample = 0.7, sampling_method = "uniform", colsample_bytree = 1, colsample_bylevel = 1, colsample_bynode = 1, lambda = 1, alpha = 0,
+                           max_leaves = 0, max_bin = 256, num_parallel_tree = 1, nthread = -1) {
+  list(
+    device = device, tree_method = tree_method, eta = eta, gamma = gamma, max_depth = max_depth, min_child_weight = min_child_weight,
+    subsample = subsample, sampling_method = sampling_method, colsample_bytree = colsample_bytree, colsample_bylevel = colsample_bylevel,
+    colsample_bynode = colsample_bynode, lambda = lambda, alpha = alpha,
+    max_leaves = max_leaves, max_bin = max_bin, num_parallel_tree = num_parallel_tree, nthread = nthread
+  )
+}
+
+
+is_dir <- function(dir) {
+  return(file.info(dir)$isdir)
 }